+++ lib/l10n/app_localizations_en.dart
@@ -64,6 +64,9 @@ class AppLocalizationsEn extends AppLocalizations {
   String get captures => 'Captures';
 
   @override
+  String get clearBoard => 'Clear';
+
+  @override
   String get collectStats => 'Collect statistics';
 
   @override
@@ -106,6 +109,9 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String get edgeLine => 'Edge line';
+
+  @override
+  String get empty => 'Empty';
 
   @override
   String get endgameExam => 'Endgame exam';
@@ -154,6 +160,18 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String get findTask => 'Find task';
+
+  @override
+  String get findTaskByLink => 'By link';
+
+  @override
+  String get findTaskByPattern => 'By pattern';
+
+  @override
+  String get findTaskResults => 'Search results';
+
+  @override
+  String get findTaskSearching => 'Searching...';
 
   @override
   String get forceCounting => 'Force counting';
@@ -496,6 +514,9 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String get taskNext => 'Next';
+
+  @override
+  String get taskNotFound => 'Task not found';
 
   @override
   String get taskRedo => 'Redo';